requirements.txt
pandas 
polars 
duckdb 
fastapi 
uvicorn 
pyyaml
python-multipart
pyarrow
openpyxl
lxml